refactor: simplify ReaderPageProgress by extracting usePage hook and removing unused atoms
