+++ apps/web/src/routes/kiwi/-reader/hooks/index.ts
@@ -3,3 +3,4 @@ export { default as useObserver } from "./use-observer";
 export { default as useRelocated } from "./use-relocated";
 export { default as useRendered } from "./use-rendered";
 export { default as useRender } from "./use-render";
+export { default as usePage } from "./use-page";
